UI: Respect the pause button display setting.
As of 22782b6, the setting stopped working.  It can be useful for Windows
touch devices.  Fixes #9609.

+++ Windows/XinputDevice.cpp
@@ -113,7 +113,7 @@ XinputDevice::XinputDevice() {
 	}
 
 	for (size_t i = 0; i < ARRAY_SIZE(check_delay); ++i) {
-		check_delay[i] = i;
+		check_delay[i] = (int)i;
 	}
 }
